requirements.txt
numpy==1.20.1
wget==3.2
requests==2.25.1
googletrans==3.1.0a0
fasttext==0.9.2
nltk==3.6.1
Keras==2.4.3
transformers==4.5.1
Pillow==8.2.0
